% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SanAntonioApp-package.R
\docType{package}
\name{SanAntonioApp-package}
\alias{SanAntonioApp}
\alias{SanAntonioApp-package}
\title{SanAntonioApp: What the Package Does (Title Case)}
\description{
Curvas de duración de caudales para la cuenca del San Antonio - Salto (Uruguay).
}
\keyword{internal}
